+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to
 - 💬(backend) clarify french transcription audio download link text #1299
 - 🚧(addons) introduce initial Microsoft Outlook add-in support (alpha)
 - 🔧(backend) add setting to toggle application token exchange mechanism
+- ✨(backend) support add-ons authentication in external viewset
 
 ### Fixed